numeric types cannot aggregate

+++ backend/recommend.py
@@ -18,7 +18,7 @@ def get_user_ratings(pair, media_id, mediaType='Movie', num=15):
     pair[1].execute("SELECT user_id, media_id, rating, watched, liked FROM preferences WHERE media_id IN (SELECT ID FROM media WHERE mediaType = %s);", (mediaType,))
 
     table = pair[1].fetchall()
-    table = [(a, b, int(c)+2*int(d)+3*int(e)) for a, b, c, d, e in table]
+    table = [(a, b, float(float(c)+2*int(d)+3*int(e))) for a, b, c, d, e in table]
 
     #list of all ratings
     df = pd.DataFrame(table, columns=["user_id", "media_id", "rating"])
